Move leader definition to Term
This makes it easier to define Terms as having leaders.

Now Terms have a set of States that map to a Node as that Term's leader.
The definition requiring every Term to have a leader needs to be updated
to make sure that Node is the leader for the rest of the Term as well.

Also the models generated by these new constraints showed that the
constraints for States in a Term are not clear enough and need to be
reworked to make sure States cannot go out of order between Terms.