Removed State from Term leader relation
A Term must have one leader Node, if/when elections are implemented an
'election Term' can possibly be a sig that extends Term.

This makes the one leader per term constraint unnecessary as it is part
of the sig and it makes the non-consecutive terms easy to implement as
well.