Add update messages
Nodes could change value at any time prior to this. This change
constrains Nodes to change value only when told to by a message from the
leader of the current Term.

Currently Nodes must update directly after receiving a message from the
leader, this should be changed to allow them to update at any time in
the current term.

As a result of these changes the StatesInOrder assertion needed to be
dropped from 40 to 10 to stay under default memory limits.